requirements.txt
absl-py==0.9.0
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blessings==1.7
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
detectron2==0.1+cu100
dominate==2.5.1
entrypoints==0.3
filecache==0.77
fire==0.2.1
future==0.18.2
fvcore==0.1.dev200218
google-auth==1.10.2
google-auth-oauthlib==0.4.1
gpustat==0.6.0
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.8.0
importlib-metadata==1.4.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mat4py==0.4.2
matplotlib==3.1.2
meshio==2.0.0
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
mpld3==0.3
munch==2.5.0
mypy==0.770
mypy-extensions==0.4.3
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
ninja==1.9.0.post1
nltk==3.4.5
notebook==6.0.3
numpy==1.17.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.2.30
pandas==1.0.0
pandocfilters==1.4.2
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
pipdate==0.2.4
portalocker==1.5.2
-e git+https://github.com/alexandonian/pretorched-x@74f92fa0ed1c64fa4bc0c4aaf86f56d315cb0c45#egg=pretorched
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.2
psutil==5.6.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pydot==1.4.1
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sru==2.3.2
tabulate==0.8.6
tensorboard==2.1.0
tensorboardX==2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
visual-genome==1.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
yacs==0.1.6
yapf==0.29.0
zipp==0.6.0
